removing old homepage and use updated one
